Add fsd_utils logging and gunicorn application startup, with updates to required dependencies using pip-tools
