+++ README.md
@@ -36,7 +36,15 @@ Built with Flask.
 ### Install dependencies
 From the top-level directory enter the command to install pip and the dependencies of the project
 
-    python3 -m pip install --upgrade pip && pip install -r requirements.txt
+    python3 -m pip install --upgrade pip && pip install -r requirements-dev.txt
+
+NOTE: requirements-dev.txt and requirements.txt are updated using [pip-tools pip-compile](https://github.com/jazzband/pip-tools)
+To update requirements please manually add the dependencies in the .in files (not the requirements.txt files)
+Then run:
+
+    pip-compile requirements.in
+
+    pip-compile requirements-dev.in
 
 ## How to use
 Enter the virtual environment as described above, then:
@@ -76,7 +84,7 @@ First set the FLASK_ENV environment you wish to test eg:
 
 Then run gunicorn using the following command:
 
-    gunicorn wsgi:app -c gunicorn.config.local.py
+    gunicorn wsgi:app -c run/gunicorn/local.py
 
 # Configuration
 
@@ -104,8 +112,6 @@ Performance tests are stored in a separate repository which is then run in the p
 
 This repo comes with a .pre-commit-config.yaml, if you wish to use this do
 the following while in your virtual enviroment:
-
-    pip install pre-commit black
 
     pre-commit install
