+++ config/__init__.py
@@ -5,25 +5,17 @@ FLASK_ENV = environ.get("FLASK_ENV")
 
 match FLASK_ENV:
     case "development":
-        from config.environments.development import (
-            DevelopmentConfig as Config,
-        )
+        from config.envs.development import DevelopmentConfig as Config
     case "dev":
-        from config.environments.dev import (
-            DevConfig as Config,
-        )
+        from config.envs.dev import DevConfig as Config
     case "test":
-        from config.environments.test import TestConfig as Config
+        from config.envs.test import TestConfig as Config
     case "unit_test":
-        from config.environments.unit_test import (
-            UnitTestConfig as Config,
-        )
+        from config.envs.unit_test import UnitTestConfig as Config
     case "production":
-        from config.environments.production import (
-            ProductionConfig as Config,
-        )
+        from config.envs.production import ProductionConfig as Config
     case _:
-        from config.environments.default import DefaultConfig as Config
+        from config.envs.default import DefaultConfig as Config
 
 try:
     Config.pretty_print()
