+++ requirements-dev.in
@@ -1,0 +1,15 @@
+-r requirements.txt
+bandit
+black
+deepdiff
+pre-commit
+pytest
+pytest-env
+pytest-flask
+pytest-mock==3.7.0
+pytest-selenium==2.0.1
+pip-tools
+axe-selenium-python==2.1.6
+json2html==1.3.0
+selenium==4.2.0
+webdriver-manager==3.7.0
