+++ requirements.in
@@ -1,0 +1,54 @@
+#-----------------------------------
+#   FSD Utils
+#-----------------------------------
+https://github.com/communitiesuk/funding-service-design-utils/archive/refs/tags/0.0.8.tar.gz
+requests
+
+#-----------------------------------
+#   Flask Version
+#-----------------------------------
+Flask==2.1.1
+
+#-----------------------------------
+#   Static Assets
+#-----------------------------------
+Flask-Assets==2.0
+Flask-Compress==1.10.1
+cssmin==0.2.0
+pyScss==1.4.0
+jsmin==3.0.1
+govuk-frontend-jinja==2.0.0
+
+#-----------------------------------
+#   Forms
+#-----------------------------------
+Flask-WTF==1.0.0
+email-validator==1.2.1
+
+#-----------------------------------
+#   Redis / Session Management
+#-----------------------------------
+flask-redis==0.4.0
+Flask-Session==0.3.2
+flask-talisman==0.8.1
+PyJWT==2.4.0
+
+#-----------------------------------
+#   Azure Active Directory
+#-----------------------------------
+msal==1.18.0
+
+#-----------------------------------
+#   Database
+#-----------------------------------
+SQLAlchemy
+Flask-SQLAlchemy
+Flask-Migrate
+
+#-----------------------------------
+#   Connexion APIs
+#-----------------------------------
+connexion
+swagger-ui-bundle
+openapi-spec-validator
+prance
