+++ requirements.txt
@@ -1,116 +1,195 @@
-async-generator==1.10
+#
+# This file is autogenerated by pip-compile with python 3.10
+# To update, run:
+#
+#    pip-compile requirements.in
+#
+alembic==1.8.0
+    # via flask-migrate
 async-timeout==4.0.2
-atomicwrites==1.4.0
+    # via redis
 attrs==21.4.0
-axe-selenium-python==2.1.6
-bandit==1.7.1
-black==21.12b0
-Brotli==1.0.9
-cachelib==0.8.0
-certifi==2022.5.18.1
-cffi==1.15.0
-cfgv==3.3.1
+    # via jsonschema
+brotli==1.0.9
+    # via flask-compress
+cachelib==0.9.0
+    # via flask-session
+certifi==2022.6.15
+    # via requests
+cffi==1.15.1
+    # via cryptography
 chardet==4.0.0
+    # via prance
 charset-normalizer==2.0.12
+    # via requests
 click==8.1.3
+    # via
+    #   clickclick
+    #   flask
 clickclick==20.10.2
-colorama==0.4.4
+    # via connexion
 commonmark==0.9.1
-connexion==2.13.1
+    # via rich
+connexion==2.14.0
+    # via -r requirements.in
 cryptography==37.0.2
+    # via
+    #   msal
+    #   pyjwt
 cssmin==0.2.0
-deepdiff==5.7.0
-Deprecated==1.2.13
-distlib==0.3.4
+    # via -r requirements.in
+deprecated==1.2.13
+    # via redis
 dnspython==2.2.1
-dparse==0.5.1
+    # via email-validator
 email-validator==1.2.1
-filelock==3.7.1
-Flask==2.0.2
-Flask-Assets==2.0
-Flask-Compress==1.10.1
+    # via -r requirements.in
+flask==2.1.1
+    # via
+    #   -r requirements.in
+    #   connexion
+    #   flask-assets
+    #   flask-compress
+    #   flask-migrate
+    #   flask-redis
+    #   flask-session
+    #   flask-sqlalchemy
+    #   flask-wtf
+    #   funding-service-design-utils
+flask-assets==2.0
+    # via -r requirements.in
+flask-compress==1.10.1
+    # via -r requirements.in
+flask-migrate==3.1.0
+    # via -r requirements.in
 flask-redis==0.4.0
-Flask-Session==0.3.2
+    # via -r requirements.in
+flask-session==0.3.2
+    # via -r requirements.in
+flask-sqlalchemy==2.5.1
+    # via
+    #   -r requirements.in
+    #   flask-migrate
 flask-talisman==0.8.1
-Flask-WTF==1.0.0
-git+https://github.com/communitiesuk/funding-service-design-utils.git@main
-gitdb==4.0.9
-GitPython==3.1.27
+    # via -r requirements.in
+flask-wtf==1.0.0
+    # via -r requirements.in
+funding-service-design-utils @ https://github.com/communitiesuk/funding-service-design-utils/archive/refs/tags/0.0.8.tar.gz
+    # via -r requirements.in
 govuk-frontend-jinja==2.0.0
+    # via -r requirements.in
+greenlet==1.1.2
+    # via sqlalchemy
 gunicorn==20.1.0
-h11==0.13.0
-identify==2.5.1
+    # via funding-service-design-utils
 idna==3.3
+    # via
+    #   email-validator
+    #   requests
 inflection==0.5.1
-iniconfig==1.1.1
+    # via connexion
 itsdangerous==2.1.2
-Jinja2==3.1.2
+    # via
+    #   connexion
+    #   flask
+    #   flask-wtf
+jinja2==3.1.2
+    # via
+    #   flask
+    #   govuk-frontend-jinja
+    #   swagger-ui-bundle
 jsmin==3.0.1
-json2html==1.3.0
+    # via -r requirements.in
 jsonschema==4.6.0
-MarkupSafe==2.0.1
+    # via
+    #   connexion
+    #   openapi-schema-validator
+    #   openapi-spec-validator
+mako==1.2.0
+    # via alembic
+markupsafe==2.1.1
+    # via
+    #   jinja2
+    #   mako
+    #   wtforms
 msal==1.18.0
-mypy-extensions==0.4.3
-nodeenv==1.6.0
+    # via -r requirements.in
 openapi-schema-validator==0.2.3
+    # via openapi-spec-validator
 openapi-spec-validator==0.4.0
-ordered-set==4.0.2
-outcome==1.2.0
+    # via -r requirements.in
 packaging==21.3
-pathspec==0.9.0
-pbr==5.9.0
-platformdirs==2.5.2
-pluggy==1.0.0
+    # via
+    #   connexion
+    #   redis
 prance==0.21.8.0
-pre-commit==2.16.0
-py==1.11.0
+    # via -r requirements.in
 pycparser==2.21
-Pygments==2.12.0
-PyJWT==2.4.0
-pyOpenSSL==22.0.0
+    # via cffi
+pygments==2.12.0
+    # via rich
+pyjwt[crypto]==2.4.0
+    # via
+    #   -r requirements.in
+    #   msal
 pyparsing==3.0.9
+    # via packaging
 pyrsistent==0.18.1
-pyScss==1.4.0
-PySocks==1.7.1
-pytest==7.0.1
-pytest-base-url==2.0.0
-pytest-env==0.6.2
-pytest-flask==1.2.0
-pytest-html==3.1.1
-pytest-metadata==2.0.1
-pytest-mock==3.7.0
-pytest-selenium==2.0.1
-pytest-variables==2.0.0
+    # via jsonschema
+pyscss==1.4.0
+    # via -r requirements.in
 python-dotenv==0.20.0
+    # via funding-service-design-utils
 python-json-logger==2.0.2
+    # via funding-service-design-utils
 pytz==2022.1
-PyYAML==6.0
-redis==4.3.3
+    # via funding-service-design-utils
+pyyaml==6.0
+    # via
+    #   clickclick
+    #   connexion
+    #   funding-service-design-utils
+    #   openapi-spec-validator
+redis==4.3.4
+    # via flask-redis
 requests==2.28.0
+    # via
+    #   -r requirements.in
+    #   connexion
+    #   msal
+    #   prance
 rich==12.4.4
-ruamel.yaml==0.17.21
-ruamel.yaml.clib==0.2.6
-safety==1.10.3
-selenium==4.2.0
+    # via funding-service-design-utils
+ruamel-yaml==0.17.21
+    # via prance
+ruamel-yaml-clib==0.2.6
+    # via ruamel-yaml
 semver==2.13.0
+    # via prance
 six==1.16.0
-smmap==5.0.0
-sniffio==1.2.0
-sortedcontainers==2.4.0
-stevedore==3.5.0
+    # via
+    #   flask-talisman
+    #   prance
+    #   pyscss
+sqlalchemy==1.4.39
+    # via
+    #   -r requirements.in
+    #   alembic
+    #   flask-sqlalchemy
 swagger-ui-bundle==0.0.9
-tenacity==6.3.1
-termcolor==1.1.0
-toml==0.10.2
-tomli==1.2.3
-trio==0.21.0
-trio-websocket==0.9.2
-typing_extensions==4.2.0
+    # via -r requirements.in
 urllib3==1.26.9
-virtualenv==20.14.1
+    # via requests
 webassets==2.0
-webdriver-manager==3.7.0
-Werkzeug==2.0.3
+    # via flask-assets
+werkzeug==2.1.2
+    # via
+    #   connexion
+    #   flask
 wrapt==1.14.1
-wsproto==1.1.0
-WTForms==3.0.1
+    # via deprecated
+wtforms==3.0.1
+    # via flask-wtf
+
+# The following packages are considered to be unsafe in a requirements file:
+# setuptools
