+++ run/gunicorn/devtest.py
@@ -1,0 +1,1 @@
+from fsd_utils.gunicorn.config.devtest import *  # noqa
